requirements.txt
torch>=2.3
transformers>=4.42
datasets>=2.19
accelerate>=0.33
numpy>=1.26
pandas>=2.2
tqdm>=4.66
scikit-learn>=1.4
hnswlib>=0.8.0
faiss-cpu>=1.7.4
FlagEmbedding>=1.2.10
pyyaml>=6.0.1
tabulate>=0.9.0
